refactor: created controllers for endpoint callback functions
