+++ src/api/controllers/userController.js
@@ -1,0 +1,35 @@
+const getProfile = async (req, res) => {
+  try {
+    res.send(req.user);
+  } catch (e) {
+    res.status(500).send();
+  }
+};
+
+const updateProfile = async (req, res) => {
+  const updates = Object.keys(req.body);
+  const allowedUpdates = ['name', 'email', 'password'];
+  const isValid = updates.every((update) => allowedUpdates.includes(update));
+
+  if (!isValid) {
+    return res.status(400).send({ error: 'Invalid update' });
+  }
+  try {
+    updates.forEach((update) => (req.user[update] = req.body[update]));
+    await req.user.save();
+    res.send(req.user);
+  } catch (e) {
+    res.status(400).send(e);
+  }
+};
+
+const deleteProfile = async (req, res) => {
+  try {
+    await req.user.remove();
+    res.send(req.user);
+  } catch (e) {
+    res.status(500).send();
+  }
+};
+
+module.exports = { getProfile, updateProfile, deleteProfile };
