--- src/api/routes/index.js
@@ -1,3 +1,0 @@
-module.exports = (app) => {
-  app.use('/api/v1', require('./v1'));
-};
