+++ src/api/routes/v1/auth.js
@@ -1,50 +1,13 @@
 const router = require('express').Router();
-const User = require('../../../models/user');
+const controller = require('../../controllers/authController');
 const auth = require('../../middleware/auth');
 
-router.post('/register', async (req, res) => {
-  const user = new User(req.body);
+router.post('/register', controller.registerUser);
 
-  try {
-    await user.save();
-    const token = await user.generateAuthToken();
-    res.status(201).send({ user, token });
-  } catch (e) {
-    res.status(400).send(e);
-  }
-});
+router.post('/login', controller.loginUser);
 
-router.post('/login', async (req, res) => {
-  try {
-    const user = await User.findByCredentials(req.body.email, req.body.password);
-    const token = await user.generateAuthToken();
-    res.send({ user, token });
-  } catch (e) {
-    res.status(400).send();
-  }
-});
+router.post('/logout', auth, controller.logoutUser);
 
-router.post('/logout', auth, async (req, res) => {
-  try {
-    req.user.tokens = req.user.tokens.filter((token) => {
-      return token.token !== req.token;
-    });
-    await req.user.save();
-
-    res.send();
-  } catch (e) {
-    res.status(500).send();
-  }
-});
-
-router.post('/logoutAll', auth, async (req, res) => {
-  try {
-    req.user.tokens = [];
-    await req.user.save();
-    res.send();
-  } catch (e) {
-    res.status(500).send();
-  }
-});
+router.post('/logoutAll', auth, controller.logoutAllSessions);
 
 module.exports = router;
