+++ src/api/routes/v1/user.js
@@ -1,38 +1,11 @@
 const router = require('express').Router();
+const controller = require('../../controllers/userController');
 const auth = require('../../middleware/auth');
 
-router.get('/profile', auth, async (req, res) => {
-  try {
-    res.send(req.user);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
+router.get('/profile', auth, controller.getProfile);
 
-router.patch('/profile', auth, async (req, res) => {
-  const updates = Object.keys(req.body);
-  const allowedUpdates = ['name', 'email', 'password'];
-  const isValid = updates.every((update) => allowedUpdates.includes(update));
+router.patch('/profile', auth, controller.updateProfile);
 
-  if (!isValid) {
-    return res.status(400).send({ error: 'Invalid update' });
-  }
-  try {
-    updates.forEach((update) => (req.user[update] = req.body[update]));
-    await req.user.save();
-    res.send(req.user);
-  } catch (e) {
-    res.status(400).send(e);
-  }
-});
-
-router.delete('/', auth, async (req, res) => {
-  try {
-    await req.user.remove();
-    res.send(req.user);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
+router.delete('/', auth, controller.deleteProfile);
 
 module.exports = router;
