+++ src/app.js
@@ -3,6 +3,6 @@ const express = require('express');
 const app = express();
 
 app.use(express.json());
-require('./api/routes')(app);
+app.use('/api/v1', require('./api/routes/v1'));
 
 module.exports = app;
