+++ src/server.js
@@ -1,11 +1,11 @@
 const app = require('./app');
-const db = require('./db/connect');
+const connectMongo = require('./db/connect');
 
 const port = process.env.PORT;
 
 (async () => {
   try {
-    await db.connectMongo();
+    await connectMongo();
 
     app.listen(port, () => {
       console.log(`Application is listening on ${port}`);
