+++ tests/routes/v1/auth.test.js
@@ -1,11 +1,11 @@
 const request = require('supertest');
 const app = require('../../../src/app');
-const db = require('../../../src/db/connect');
+const connectMongo = require('../../../src/db/connect');
 const User = require('../../../src/models/user');
 const { userOne, userOneId, setupDatabase, disconnectMongo } = require('../../fixtures/db');
 
 beforeAll((done) => {
-  db.connectMongo();
+  connectMongo();
   done();
 });
