+++ tests/routes/v1/inspiration.test.js
@@ -1,6 +1,6 @@
 const request = require('supertest');
 const app = require('../../../src/app');
-const db = require('../../../src/db/connect');
+const connectMongo = require('../../../src/db/connect');
 const Inspiration = require('../../../src/models/inspiration');
 const {
   userOne,
@@ -16,7 +16,7 @@ const {
 } = require('../../fixtures/db');
 
 beforeAll((done) => {
-  db.connectMongo();
+  connectMongo();
   done();
 });
