+++ tests/routes/v1/user.test.js
@@ -1,12 +1,12 @@
 const request = require('supertest');
 const app = require('../../../src/app');
-const db = require('../../../src/db/connect');
+const connectMongo = require('../../../src/db/connect');
 const User = require('../../../src/models/user');
 const Character = require('../../../src/models/character');
 const { userOne, userOneId, setupDatabase, disconnectMongo } = require('../../fixtures/db');
 
 beforeAll((done) => {
-  db.connectMongo();
+  connectMongo();
   done();
 });
